app/requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
requests==2.25.1
gunicorn==20.0.4
dj-database-url==0.5.0
djangorestframework-simplejwt==4.6.0